Accept 'top' as an attribute on <dmrs> elements
PyDelphin v1.0 uses the newer convention of putting 'top' as an
attribute on <dmrs> in the DMRX codec, but the old method of using a
link from the special nodeid 0 is still supported on reading.

This commit does not make pydmrs's dumps_xml function write DMRXs with
'top' attributes, but only allows it to read them.